글로벌 클래스 수정 및 Product 관련 수정

+++ src/main/java/com/market/carrot/product/service/ProductServiceImpl.java
@@ -4,6 +4,8 @@ import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
 
 import com.market.carrot.category.domain.Category;
 import com.market.carrot.category.domain.CategoryRepository;
+import com.market.carrot.global.Exception.ExceptionMessage;
+import com.market.carrot.global.Exception.IsNotWriterException;
 import com.market.carrot.global.Exception.NotFoundEntityException;
 import com.market.carrot.login.config.customAuthentication.common.MemberContext;
 import com.market.carrot.login.domain.Member;
@@ -76,7 +78,7 @@ public class ProductServiceImpl implements ProductService {
   @Override
   public ProductModel detail(Long id, MemberContext member) {
     Product findProduct = productRepository.findById(id).orElseThrow(
-        () -> new NotFoundEntityException("해당 제품이 존재하지 않습니다.", HttpStatus.BAD_REQUEST));
+        () -> new NotFoundEntityException(ExceptionMessage.NOT_FOUND_PRODUCT, HttpStatus.BAD_REQUEST));
 
     ProductResponse productResponse = ProductResponse.builder()
         .id(findProduct.getId())
@@ -106,10 +108,10 @@ public class ProductServiceImpl implements ProductService {
     Product saveProduct = Product.createProduct(title, content, price);
 
     Member findMember = memberRepository.findById(member.getMember().getId())
-        .orElseThrow(() -> new NotFoundEntityException("찾을 수 없는 회원입니다.", HttpStatus.BAD_REQUEST));
+        .orElseThrow(() -> new NotFoundEntityException(ExceptionMessage.NOT_FOUND_MEMBER, HttpStatus.BAD_REQUEST));
 
     Category findCategory = categoryRepository.findById(categoryId)
-        .orElseThrow(() -> new NotFoundEntityException("찾을 수 없는 카테고리입니다.", HttpStatus.BAD_REQUEST));
+        .orElseThrow(() -> new NotFoundEntityException(ExceptionMessage.NOT_FOUND_CATEGORY, HttpStatus.BAD_REQUEST));
 
     // 연관관계에 의한 Member, Category, Image 값 셋팅
     saveProduct.addMember(findMember);
@@ -131,19 +133,30 @@ public class ProductServiceImpl implements ProductService {
 
   @Transactional
   @Override
-  public void update(Long id, UpdateProductRequest productRequest) {
+  public void update(Long id, UpdateProductRequest productRequest, MemberContext member) {
     Product findProduct = productRepository.findById(id)
-        .orElseThrow(() -> new NotFoundEntityException("찾을 수 없는 제품입니다.", HttpStatus.BAD_REQUEST));
-    findProduct.updateProduct(productRequest);
+        .orElseThrow(() -> new NotFoundEntityException(ExceptionMessage.NOT_FOUND_PRODUCT, HttpStatus.BAD_REQUEST));
+
+    Member findMember = memberRepository.findById(member.getMember().getId())
+        .orElseThrow(() -> new NotFoundEntityException(ExceptionMessage.NOT_FOUND_MEMBER, HttpStatus.BAD_REQUEST));
+
+    findProduct.updateProduct(productRequest, findMember);
   }
 
   @Transactional
   @Override
-  public void delete(Long id) {
+  public void delete(Long id, MemberContext member) {
     Product findProduct = productRepository.findById(id)
-        .orElseThrow(() -> new NotFoundEntityException("찾을 수 없는 제품입니다.", HttpStatus.BAD_REQUEST));
+        .orElseThrow(() -> new NotFoundEntityException(ExceptionMessage.NOT_FOUND_PRODUCT, HttpStatus.BAD_REQUEST));
 
-    productRepository.delete(findProduct);
+    Member findMember = memberRepository.findById(member.getMember().getId())
+        .orElseThrow(() -> new NotFoundEntityException(ExceptionMessage.NOT_FOUND_MEMBER, HttpStatus.BAD_REQUEST));
+
+    if (findProduct.checkUser(findMember)) {
+      productRepository.delete(findProduct);
+    }
+
+    throw new IsNotWriterException(ExceptionMessage.IS_NOT_WRITER_BY_DELETE, HttpStatus.BAD_REQUEST);
   }
 
   /**
@@ -156,13 +169,13 @@ public class ProductServiceImpl implements ProductService {
   private ProductModel getResponseByProductDetail(MemberContext member,
       ProductResponse productResponse) {
     Long productId = productResponse.getId();
-    Long memberOfProductId = productResponse.getMember().getId();
+    String memberOfProductName = productResponse.getMember().getUsername();
 
     ProductModel entityModelByProductResponse = new ProductModel(productResponse);
     addHateoasLink(entityModelByProductResponse, productId);
 
     // 자신이 작성한 상품인 경우 삭제 및 수정 API 호출이 가능하다.
-    if (member != null && member.getMember().getId().equals(memberOfProductId)) {
+    if (member != null && member.getMember().getUsername().equals(memberOfProductName)) {
       addHateoasLink(entityModelByProductResponse, productId, "product-delete");
       addHateoasLink(entityModelByProductResponse, productId, "product-update");
     }
